[demux src] merge the tow kind of objects

+++ src/src_curl.c
@@ -198,7 +198,7 @@ static void src_eventlistener(src_ctx_t *ctx, event_listener_cb_t cb, void *arg)
 	}
 }
 
-static int src_attach(src_ctx_t *ctx, int index, decoder_t *decoder)
+static int src_attach(src_ctx_t *ctx, long index, decoder_t *decoder)
 {
 	if (index > 0)
 		return -1;
@@ -209,7 +209,7 @@ static int src_attach(src_ctx_t *ctx, int index, decoder_t *decoder)
 				(curl_off_t)ctx->out->ctx->size * ctx->out->ctx->count);
 }
 
-static decoder_t *src_estream(src_ctx_t *ctx, int index)
+static decoder_t *src_estream(src_ctx_t *ctx, long index)
 {
 	return ctx->estream;
 }
