+++ src/src_unix.c
@@ -237,7 +237,7 @@ static void src_eventlistener(src_ctx_t *ctx, event_listener_cb_t cb, void *arg)
 	}
 }
 
-static int src_attach(src_ctx_t *ctx, int index, decoder_t *decoder)
+static int src_attach(src_ctx_t *ctx, long index, decoder_t *decoder)
 {
 	if (index > 0)
 		return -1;
@@ -245,7 +245,7 @@ static int src_attach(src_ctx_t *ctx, int index, decoder_t *decoder)
 	ctx->out = ctx->estream->ops->jitter(ctx->estream->ctx, JITTE_MID);
 }
 
-static decoder_t *src_estream(src_ctx_t *ctx, int index)
+static decoder_t *src_estream(src_ctx_t *ctx, long index)
 {
 	return ctx->estream;
 }
